Refactored transformer by splitting it into multiple transformers
